requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-common-sql